pyproject.toml
[build-system]
requires = [
    "setuptools >=80.0.0, <81.0.0", 
    "wheel >=0.45.0, <0.46.0",
]
build-backend = "setuptools.build_meta"

[project]
name = "delivery"
version = "0.1.0"
requires-python = ">=3.13, <3.14"
readme = "README.md"
dependencies = [
    "fastapi==0.119.1",
    "Hypercorn==0.17.3",
    "pydantic==2.12.3",
    "SQLAlchemy==2.0.44",
    "aiosqlite==0.21.0",
    "coloredlogs==15.0.1",
    # "PyYAML==6.0.3",
    "pika==1.3.2",
    "PyJWT==2.10.1",
    "chassis @ git+https://github.com/monolith2microservice/Chassis.git"
]

[project.optional-dependencies]
dev = [
    "build==1.3.0",
]

[project.scripts]
delivery = "delivery:start_server"

[tool.setuptools.package-data]
machine = [
    "logging.ini",
    ".pylintrc",
]